refactor(pkg): rename pkg folder to lib
