+++ lib/logger.go
@@ -1,0 +1,1 @@
+package lib
